Updated validation logic
